Move style files into components' folder
