+++ src/components/Sidebar/SidebarMain/SidebarMainSection/SidebarMainSectionDefault.js
@@ -1,10 +1,10 @@
-import sidebarMainSectionStyle from '../../../../styles/components/sidebar/sidebarMain/sidebarMainSectionStyle.js';
+import SidebarMainSection from '../SidebarMainSection.js';
+import sidebarMainSectionStyle from './sidebarMainSectionStyle';
 import {
   SIDEBAR_CATEGORY_HEIGHT,
   SIDEBAR_FOLDING_BTN_HEIGHT,
   FOLD_THRESHOLD,
-} from '../../../../constant.js';
-import SidebarMainSection from '../SidebarMainSection.js';
+} from '@constant';
 
 class SidebarMainSectionDefault extends SidebarMainSection {
   constructor() {
