+++ src/components/Sidebar/SidebarMain/sidebarFoldStyle.js
@@ -1,4 +1,4 @@
-import { SIDEBAR_FOLDING_BTN_HEIGHT } from '../../../../constant.js';
+import { SIDEBAR_FOLDING_BTN_HEIGHT } from '../../../constant.js';
 
 export default function sidebarFoldStyle() {
   const style = document.createElement('style');
