+++ src/components/iconStyle.js
