+++ src/scripts/sidebar.js
@@ -1,7 +1,7 @@
-import Sidebar from '../components/Sidebar.js';
+import Sidebar from '@components/Sidebar.js';
 import { loginActionElement } from './action.js';
-import SidebarSubContent from '../components/Sidebar/SidebarSub/SidebarSubContent.js';
-import { menus } from '../data/menu.js';
+import SidebarSubContent from '@components/Sidebar/SidebarSub/SidebarSubContent.js';
+import { menus } from '@data/menu.js';
 
 const showAllBtn = document
   .querySelector('navbar-element')
