+++ vite.config.js
@@ -1,0 +1,13 @@
+export default {
+  resolve: {
+    alias: {
+      '@src': '/src',
+      '@components': '/src/components',
+      '@style': '/src/style',
+      '@scripts': '/src/scripts',
+      '@constant': '/src/constant.js',
+      '@data': '/src/data',
+      '@assets': '/public/assets',
+    },
+  },
+};
